+++ src_plugins/ppx_deriving_ord.ml
@@ -139,6 +139,7 @@ let () =
       [Sig.value (Val.mk (mknoloc ("compare_"^type_.ptype_name.txt))
                   (polymorphize [%type: [%t typ] -> [%t typ] -> int]))]
     in
-    [Str.value Recursive (List.concat (List.map expr_of_type type_decls))],
+    Ppx_deriving.catch (fun () ->
+      [Str.value Recursive (List.concat (List.map expr_of_type type_decls))]),
     List.concat (List.map sig_of_type type_decls))