+++ src_plugins/ppx_deriving_show.ml
@@ -130,5 +130,6 @@ let () =
        Sig.value (Val.mk (mknoloc ("show_"^type_.ptype_name.txt))
                   (polymorphize [%type: [%t typ] -> string]))]
     in
-    [Str.value Recursive (List.concat (List.map expr_of_type type_decls))],
+    Ppx_deriving.catch (fun () ->
+      [Str.value Recursive (List.concat (List.map expr_of_type type_decls))]),
     List.concat (List.map sig_of_type type_decls))