+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 0.2
 ---
 
-  * Added Ppx_deriving.expand_path.
+  * Added Ppx_deriving.{catch,expand_path}.
   * Change Ppx_deriving.deriver signature to pass ~options and ~path as labeled arguments.
   * Show: include full module path in output.
   * Show, Eq, Ord: add support for list, array, option.